refactor(sinks): ynab sink split out transformYNAB() into transform.ts
