+++ sink/ynab/tsconfig.json
@@ -28,7 +28,7 @@
 
     /* Modules */
     // target es2022 for top level await
-    "module": "ES2022" /* Specify what module code is generated. */,
+    "module": "es2022" /* Specify what module code is generated. */,
     "rootDir": "src" /* Specify the root folder within your source files. */,
     "moduleResolution": "node" /* Specify how TypeScript looks up a file from a given module specifier. */,
     // "baseUrl": "./",                                  /* Specify the base directory to resolve non-relative module names. */
